升级SDK,处理兼容问题

+++ baidu-map-compose/src/main/java/com/melody/map/baidu_compose/overlay/TraceOverlay.kt
@@ -80,8 +80,10 @@ fun TraceOverlay(
     ComposeNode<TraceOverlayNode, MapApplier>(
         factory = {
             val listener = object : TraceAnimationListener {
-                override fun onTraceAnimationUpdate(p0: Int) {
+
+                override fun onTraceAnimationUpdate(p0: Float) {
                 }
+
                 override fun onTraceUpdatePosition(p0: LatLng?) {
                 }
                 override fun onTraceAnimationFinish() {
